requirements.txt
ffmpeg-python
git+https://github.com/openai/whisper.git
